+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [1.9.8] - unreleased
 ### Added
+- Examples command; thanks @MaxG87
 ### Changed
 ### Deprecated
 ### Removed